Update master with 2nd block challenges (#6)
This commit includes the project files for the 2nd block of challenges
with the exception of sensitive key information. Intial foundation of
recipe search app has been complete.

* added animations for recipe project

This commit includes animations for all the recipe results as well
as the heading for the project. Also fixed link errors for images.

* added styling and error handling

This commit includes card styling for the recipe results and color
to the header and body. Also added error message if users query
does not return results.

* added shadow animations / more styling

This commit adds needed styling to the header including a card background
for the instructions and a shadow animation because those are cool.

* Added unit tests and npm script

This commit adds more robust unit tests for checking the
recipe result formatting. Also added simple bash script
for experimenting how to run scripts using npm.